+++ sql/00_params.sql
@@ -879,6 +879,7 @@ BEGIN
         || '<b>SQL</b>'
         || '<a href="#topsql">Top SQL</a>'
         || '<a href="#topsql-ash">Top SQL &mdash; ASH</a>'
+        || '<a href="#sqlmon">SQL Monitor</a>'
         || '<b>Storage &amp; config</b>'
         || '<a href="#segment-io">Segment I/O</a>'
         || '<a href="#file-io">File I/O</a>'